Directories & Files reorganization
