+++ README.md
@@ -10,3 +10,10 @@ All sections are located in the GitHub repo's [Wiki](https://github.com/Will-Jav
 * [Description](https://github.com/Will-Java-FS/Team5-Project1-SarahAnatoli-StephenBecker-BolaGhaly-MichaelNugent/wiki/Project-Description)
 * [User Stories](https://github.com/Will-Java-FS/Team5-Project1-SarahAnatoli-StephenBecker-BolaGhaly-MichaelNugent/wiki/User-Stories)
 * [Database Tables](https://github.com/Will-Java-FS/Team5-Project1-SarahAnatoli-StephenBecker-BolaGhaly-MichaelNugent/wiki/Database-Tables)
+
+## API Demonstration
+After you clone this repository, visit our
+[Postman Public Workspace](https://www.postman.com/nugentmichael/workspace/team-5-project-1-calorie-tracker-app)
+to view a presassembled list of requests and easily try a few.
+Make sure you're on the desktop application version or use the Desktop agent as the browser version
+alone will not query a localhost domain.
