+++ src/main/java/com/revature/calorietracker/models/auth/TokenType.java
@@ -1,0 +1,5 @@
+package com.revature.calorietracker.models.auth;
+
+public enum TokenType {
+    BEARER
+}
