+++ src/main/java/com/revature/calorietracker/security/config/SecurityConfig.java
@@ -4,6 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.authentication.AuthenticationProvider;
+import org.springframework.security.config.Customizer;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
@@ -28,10 +29,11 @@ public class SecurityConfig {
                 .authorizeHttpRequests((req)->
                         req.requestMatchers("auth/**")
                                 .permitAll()
-                                .requestMatchers("/users/**").hasRole("ADMIN")
+//                                .requestMatchers("/users/**").hasRole("USER")
                                 .anyRequest()
                                 .authenticated()
                 )
+//                .formLogin(Customizer.withDefaults())
                 .sessionManagement(session->session.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
                 .authenticationProvider(authenticationProvider)
                 .addFilterBefore(jwtAuthenticationFilter, UsernamePasswordAuthenticationFilter.class)
