--- src/main/java/com/revature/calorietracker/security/token/TokenType.java
@@ -1,5 +1,0 @@
-package com.revature.calorietracker.security.token;
-
-public enum TokenType {
-    BEARER
-}
